// src/types/enums.ts
export enum MODE_ENUM {
  ONLINE = "Online",
  OFFLINE = "Offline",
  ONLINEANDOFFLINE = "Online and Offline ",
}

export enum TABLE_A_FIELDS {
  id = "id",
  name = "name",
  image = "image",
  district = "district",
  type = "type",
  phone = "phone",
  email = "email",
  otherContactDetails = "otherContactDetails",
  otherLinks = "otherLinks",
  mode = "mode",
  specialisation = "specialisation",
  fee = "fee",
}

export enum DISTRICT_ENUM {
  ALA = "Alappuzha",
  ERN = "Ernakulam",
  IDU = "Idukki",
  KAN = "Kannur",
  KAS = "Kasaragod",
  KOL = "Kollam",
  KOT = "Kottayam",
  KOZ = "Kozhikode",
  MAL = "Malappuram",
  PAL = "Palakkad",
  PAT = "Pathanamthitta",
  THI = "Thiruvananthapuram",
  THR = "Thrissur",
  WAY = "Wayanad",
}

export enum SPECIALISATION_ENUM {
  CBT = "CBT",
  FamilyTherapy = "Family Therapy",
  CouplesTherapy = "Couples Therapy",
  Counseling = "Counseling",
  MentalIllnessesAndDisorders = "Mental illnesses and Disorders",
  Inquiry = "Inquiry",
  QueerAffirmative = "Queer Affirmative",
  AversionTherapyForAddiction = "Aversion therapy for addiction",
  TransactionalAnalysis = "Transactional Analysis",
  AdolescenceCounseling = "Adolescence counseling",
  Anxiety = "Anxiety",
}